refactor(push): keep FcmProperties purely configurable; derive reads instead of hand-writing them
Review feedback. FCM protocol constants (4096 payload limit, envelope headroom)
move to FcmPushSender next to the 500-token cap — the properties class now holds
only configurable knobs. findByUserId on both repositories becomes a derived
query (Layer-1-scoped via the query() proxy, per the tenant guide); the custom
fragments keep only what derivation can't express: the race-safe upserts and the
bulk token deletes.
