+++ openframe-data-mongo-sync/src/main/java/com/openframe/data/repository/push/CustomPushDeviceRepository.java
@@ -1,17 +1,13 @@
 package com.openframe.data.repository.push;
 
-import com.openframe.data.document.push.PushDevice;
 import com.openframe.data.document.push.PushPlatform;
 
 import java.util.Collection;
-import java.util.List;
 
 public interface CustomPushDeviceRepository {
 
     /** Upserts by token, re-associating an existing one to the caller. @return true when a new row was created. */
     boolean registerToken(String userId, String token, PushPlatform platform);
-
-    List<PushDevice> findByUserId(String userId);
 
     boolean removeToken(String token);
