+++ openframe-data-mongo-sync/src/main/java/com/openframe/data/repository/push/PushDeviceRepository.java
@@ -4,7 +4,11 @@ import com.openframe.data.document.push.PushDevice;
 import com.openframe.data.repository.TenantAwareRepository;
 import org.springframework.data.repository.Repository;
 
+import java.util.List;
+
 /** Bare {@link Repository}, not {@code MongoRepository}: the inherited {@code count()} is not tenant-scoped. */
 @TenantAwareRepository
 public interface PushDeviceRepository extends Repository<PushDevice, String>, CustomPushDeviceRepository {
+
+    List<PushDevice> findByUserId(String userId);
 }
