+++ openframe-notification-push/src/main/java/com/openframe/notification/push/FcmProperties.java
@@ -10,10 +10,6 @@ import java.time.Duration;
 @ConfigurationProperties(prefix = "openframe.push.fcm")
 public class FcmProperties implements InitializingBean {
 
-    static final int FCM_PAYLOAD_LIMIT_BYTES = 4096;
-
-    private static final int ENVELOPE_HEADROOM_BYTES = 512;
-
     private String projectId;
 
     private int maxTitleBytes = 200;
@@ -44,14 +40,15 @@ public class FcmProperties implements InitializingBean {
                     maxTitleBytes, maxBodyBytes, maxContextBytes));
         }
 
-        long worstCase = (long) maxTitleBytes + maxBodyBytes + maxContextBytes + ENVELOPE_HEADROOM_BYTES;
-        if (worstCase > FCM_PAYLOAD_LIMIT_BYTES) {
+        long worstCase = (long) maxTitleBytes + maxBodyBytes + maxContextBytes
+                + FcmPushSender.ENVELOPE_HEADROOM_BYTES;
+        if (worstCase > FcmPushSender.FCM_PAYLOAD_LIMIT_BYTES) {
             throw new IllegalStateException(String.format(
                     "openframe.push.fcm budgets do not fit FCM's %d-byte limit: title(%d) + body(%d) "
                             + "+ context(%d) + envelope(%d) = %d. Lower one — otherwise every push is "
                             + "rejected with INVALID_ARGUMENT.",
-                    FCM_PAYLOAD_LIMIT_BYTES, maxTitleBytes, maxBodyBytes, maxContextBytes,
-                    ENVELOPE_HEADROOM_BYTES, worstCase));
+                    FcmPushSender.FCM_PAYLOAD_LIMIT_BYTES, maxTitleBytes, maxBodyBytes, maxContextBytes,
+                    FcmPushSender.ENVELOPE_HEADROOM_BYTES, worstCase));
         }
     }
 }
